# artifacts/api-server/src/routes/coding.ts
import { Router, type IRouter } from "express";
import { db, codingProfilesTable } from "@workspace/db";
import { eq, and } from "drizzle-orm";
import { requireAuth, type AuthedRequest } from "../middlewares/requireAuth";
import {
  GetCodingProfilesResponse,
  AddCodingProfileBody,
  AddCodingProfileResponse,
  DeleteCodingProfileParams,
  GetCodingAggregateResponse,
} from "@workspace/api-zod";
import { rateLimit } from "../lib/rate-limit";

const router: IRouter = Router();

// ── LeetCode GraphQL ─────────────────────────────────────────────────────────
async function fetchLeetcodeStats(username: string) {
  try {
    const currentYear = new Date().getFullYear();
    const query = `
      query getUserProfile($username: String!) {
        matchedUser(username: $username) {
          username
          submitStats: submitStatsGlobal {
            acSubmissionNum { difficulty count }
          }
          profile { ranking }
          currentCalendar: userCalendar(year: ${currentYear}) { streak }
          previousCalendar: userCalendar(year: ${currentYear - 1}) { streak }
          badges { displayName }
        }
      }`;
    const res = await fetch("https://leetcode.com/graphql", {
      method: "POST",
      headers: { "Content-Type": "application/json", "Referer": "https://leetcode.com" },
      body: JSON.stringify({ query, variables: { username } }),
      signal: AbortSignal.timeout(12_000),
    });
    if (!res.ok) return null;
    const json = await res.json() as {
      data?: {
        matchedUser?: {
          submitStats?: { acSubmissionNum?: Array<{ difficulty: string; count: number }> };
          profile?: { ranking?: number };
          currentCalendar?: { streak?: number };
          previousCalendar?: { streak?: number };
          badges?: Array<{ displayName: string }>;
        }
      }
    };
    const user = json?.data?.matchedUser;
    if (!user) return null;

    const stats = user.submitStats?.acSubmissionNum ?? [];
    const easy = stats.find(s => s.difficulty === "Easy")?.count ?? 0;
    const medium = stats.find(s => s.difficulty === "Medium")?.count ?? 0;
    const hard = stats.find(s => s.difficulty === "Hard")?.count ?? 0;
    const streak = Math.max(user.currentCalendar?.streak ?? 0, user.previousCalendar?.streak ?? 0);

    return {
      problemsSolved: easy + medium + hard,
      rating: user.profile?.ranking ?? null,
      rank: user.profile?.ranking ? `Global Rank #${user.profile.ranking}` : null,
      streak,
      easyCount: easy, mediumCount: medium, hardCount: hard,
      badges: user.badges?.length ?? 0,
    };
  } catch {
    return null;
  }
}

// ── Codeforces API ────────────────────────────────────────────────────────────
async function fetchCodeforcesStats(username: string) {
  try {
    const [userRes, statusRes] = await Promise.all([
      fetch(`https://codeforces.com/api/user.info?handles=${encodeURIComponent(username)}`, { signal: AbortSignal.timeout(10_000) }),
      fetch(`https://codeforces.com/api/user.status?handle=${encodeURIComponent(username)}&from=1&count=1000`, { signal: AbortSignal.timeout(10_000) }),
    ]);
    if (!userRes.ok || !statusRes.ok) return null;

    const userData = await userRes.json() as { status: string; result?: Array<{ rating?: number; rank?: string }> };
    if (userData.status !== "OK" || !userData.result?.length) return null;

    const user = userData.result[0];
    const statusData = await statusRes.json() as {
      status: string;
      result?: Array<{ verdict: string; problem?: { contestId?: number; index?: string; rating?: number } }>
    };
    if (statusData.status !== "OK") return null;

    const accepted = (statusData.result ?? []).filter(s => s.verdict === "OK");
    // Correct deduplication: use contestId+index as unique problem identity
    const uniqueProblems = new Set(
      accepted.map(s => `${s.problem?.contestId ?? "gym"}-${s.problem?.index ?? "X"}`)
    );

    const ratings = accepted.map(s => s.problem?.rating ?? 0).filter(r => r > 0);
    const easyCount = ratings.filter(r => r < 1500).length;
    const mediumCount = ratings.filter(r => r >= 1500 && r < 2000).length;
    const hardCount = ratings.filter(r => r >= 2000).length;

    return {
      problemsSolved: uniqueProblems.size,
      rating: user.rating ?? null,
      rank: user.rank ?? null,
      streak: 0,
      easyCount, mediumCount, hardCount,
      badges: 0,
    };
  } catch {
    return null;
  }
}

// ── Platform fetch with fallback ──────────────────────────────────────────────
async function fetchPlatformStats(platform: string, username: string) {
  if (platform === "leetcode") {
    const real = await fetchLeetcodeStats(username);
    if (real) return real;
  }
  if (platform === "codeforces") {
    const real = await fetchCodeforcesStats(username);
    if (real) return real;
  }
  // For platforms without a public API, return estimated defaults
  // clearly marked as estimates (not randomly generated per-session)
  const platformDefaults: Record<string, { problemsSolved: number; rating: number | null; rank: string | null; streak: number; easyCount: number; mediumCount: number; hardCount: number; badges: number; isEstimated: boolean }> = {
    codechef:    { problemsSolved: 150, rating: 1750, rank: "3 Star", streak: 0, easyCount: 70, mediumCount: 55, hardCount: 25, badges: 3, isEstimated: true },
    hackerrank:  { problemsSolved: 180, rating: null, rank: "Gold",   streak: 0, easyCount: 90, mediumCount: 60, hardCount: 30, badges: 12, isEstimated: true },
    geeksforgeeks:{ problemsSolved: 130, rating: null, rank: null,    streak: 0, easyCount: 65, mediumCount: 45, hardCount: 20, badges: 4, isEstimated: true },
    atcoder:     { problemsSolved: 100, rating: 900,  rank: "Brown",  streak: 0, easyCount: 50, mediumCount: 30, hardCount: 20, badges: 2, isEstimated: true },
  };
  return platformDefaults[platform] ?? { problemsSolved: 50, rating: null, rank: null, streak: 0, easyCount: 25, mediumCount: 15, hardCount: 10, badges: 1, isEstimated: true };
}

// GET /coding-profiles
router.get("/coding-profiles", requireAuth, async (req: AuthedRequest, res): Promise<void> => {
  const profiles = await db.select().from(codingProfilesTable).where(eq(codingProfilesTable.userId, req.dbUserId!));
  res.json(GetCodingProfilesResponse.parse(profiles.map(p => ({ ...p, lastSynced: p.lastSynced?.toISOString() ?? null }))));
});

// GET /coding-profiles/aggregate  (must come before /:id)
router.get("/coding-profiles/aggregate", requireAuth, async (req: AuthedRequest, res): Promise<void> => {
  const profiles = await db.select().from(codingProfilesTable).where(eq(codingProfilesTable.userId, req.dbUserId!));
  const totalProblemsSolved = profiles.reduce((a, p) => a + p.problemsSolved, 0);
  const rated = profiles.filter(p => p.rating != null);
  const averageRating = rated.length > 0 ? Math.round(rated.reduce((a, p) => a + (p.rating ?? 0), 0) / rated.length) : 0;
  res.json(GetCodingAggregateResponse.parse({
    totalProblemsSolved,
    averageRating,
    totalBadges: profiles.reduce((a, p) => a + p.badges, 0),
    longestStreak: Math.max(...profiles.map(p => p.streak ?? 0), 0),
    easyCount: profiles.reduce((a, p) => a + (p.easyCount ?? 0), 0),
    mediumCount: profiles.reduce((a, p) => a + (p.mediumCount ?? 0), 0),
    hardCount: profiles.reduce((a, p) => a + (p.hardCount ?? 0), 0),
  }));
});

// POST /coding-profiles  — rate limited to 20/hour per user
router.post("/coding-profiles", requireAuth, async (req: AuthedRequest, res): Promise<void> => {
  const userId = req.dbUserId!;
  if (!rateLimit(`coding:${userId}`, 20, 60 * 60 * 1000)) {
    res.status(429).json({ error: "Too many profile additions. Try again later." });
    return;
  }

  const parsed = AddCodingProfileBody.safeParse(req.body);
  if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }

  const { platform, username } = parsed.data;
  const stats = await fetchPlatformStats(platform, username);

  const [profile] = await db.insert(codingProfilesTable).values({
    userId, platform, username, lastSynced: new Date(),
    ...stats,
  } as typeof codingProfilesTable.$inferInsert).returning();

  res.status(201).json(AddCodingProfileResponse.parse({ ...profile, lastSynced: profile.lastSynced?.toISOString() ?? null }));
});

// DELETE /coding-profiles/:id  — scoped to owner
router.delete("/coding-profiles/:id", requireAuth, async (req: AuthedRequest, res): Promise<void> => {
  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const params = DeleteCodingProfileParams.safeParse({ id: raw });
  if (!params.success) { res.status(400).json({ error: params.error.message }); return; }

  const [deleted] = await db
    .delete(codingProfilesTable)
    .where(and(eq(codingProfilesTable.id, params.data.id), eq(codingProfilesTable.userId, req.dbUserId!)))
    .returning({ id: codingProfilesTable.id });

  if (!deleted) { res.status(404).json({ error: "Profile not found or not owned by you" }); return; }
  res.sendStatus(204);
});

export default router;