# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SkillScope – AI-Powered Developer Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /coding-profiles:
    get:
      operationId: getCodingProfiles
      tags: [coding-profiles]
      summary: List all connected competitive programming profiles
      responses:
        "200":
          description: Coding profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CodingProfile"
    post:
      operationId: addCodingProfile
      tags: [coding-profiles]
      summary: Add a new competitive programming profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodingProfileInput"
      responses:
        "201":
          description: Profile added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodingProfile"
  /coding-profiles/{id}:
    delete:
      operationId: deleteCodingProfile
      tags: [coding-profiles]
      summary: Remove a connected coding profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /career-coach/messages:
    get:
      operationId: getCareerCoachMessages
      tags: [career-coach]
      summary: Get chat history with AI career coach
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendCareerCoachMessage
      tags: [career-coach]
      summary: Send a message to the AI career coach
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
...
components:
  schemas:
...
    CodingProfile:
      type: object
      required: [id, platform, username, problemsSolved, rating, rank]
      properties:
        id:
          type: integer
        platform:
          type: string
          enum: [leetcode, codeforces, codechef, hackerrank, geeksforgeeks, atcoder]
        username:
          type: string
        problemsSolved:
          type: integer
        rating:
          type: ["integer", "null"]
        rank:
          type: ["string", "null"]
        streak:
          type: ["integer", "null"]
        easyCount:
          type: ["integer", "null"]
        mediumCount:
          type: ["integer", "null"]
        hardCount:
          type: ["integer", "null"]
        badges:
          type: integer
        lastSynced:
          type: ["string", "null"]
    CodingProfileInput:
      type: object
      required: [platform, username]
      properties:
        platform:
          type: string
          enum: [leetcode, codeforces, codechef, hackerrank, geeksforgeeks, atcoder]
        username:
          type: string
...
    ChatMessage:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ─── Roadmap ────────────────────────────────────────────────────────────────
...